+++ projects/igniteui-angular/migrations/update-18_0_0/index.spec.ts
@@ -193,4 +193,36 @@ describe(`Update to ${version}`, () => {
             <igx-row-island/>
         </igx-hierarchical-grid>`);
     });
+    
+    it('should replace PivotGrid property `showPivotConfigurationUI` with `pivotUI`', async () => {
+        appTree.create(`/testSrc/appPrefix/component/test.component.html`,
+        `
+        <igx-pivot-grid [showPivotConfigurationUI]="false"></igx-pivot-grid>
+        `
+        );
+
+        const tree = await schematicRunner.runSchematic(migrationName, { shouldInvokeLS: false }, appTree);
+
+        expect(tree.readContent('/testSrc/appPrefix/component/test.component.html')).toEqual(
+        `
+        <igx-pivot-grid [pivotUI]="{ showConfiguration: false }"></igx-pivot-grid>
+        `
+        );
+    });
+
+    it('should replace PivotGrid property `showPivotConfigurationUI` with `pivotUI`', async () => {
+        appTree.create(`/testSrc/appPrefix/component/test.component.html`,
+        `
+        <igx-pivot-grid [showPivotConfigurationUI]="testProp"></igx-pivot-grid>
+        `
+        );
+
+        const tree = await schematicRunner.runSchematic(migrationName, { shouldInvokeLS: false }, appTree);
+
+        expect(tree.readContent('/testSrc/appPrefix/component/test.component.html')).toEqual(
+        `
+        <igx-pivot-grid [pivotUI]="{ showConfiguration: testProp }"></igx-pivot-grid>
+        `
+        );
+    });
 });